refactor(cli): extract agents and packages command modules
One commit for both groups: they share AgentsCmdFlags, parseAgentsFlags,
and printPackageActionTui (exported from the agents module).
