+++ cli/bakin.ts
@@ -43,7 +43,6 @@ import {
   exitUnknownSubcommand,
   exitCommandFailure,
   promptYesNo,
-  confirmPrompt,
 } from '../src/cli/help'
 import { parsePluginInstallArgs, PLUGIN_INSTALL_USAGE } from '../src/core/cli/plugin-install-args'
 import {
@@ -56,7 +55,6 @@ import {
 import type { Permission } from '@bakin/core/plugins/permissions'
 import { renderInkReport } from '../src/core/cli/ui/render-report'
 import type {
-  PackageActionData,
   PluginActionData,
   PluginRestoreResultData,
   PluginRestoreSnapshotData,
@@ -74,18 +72,6 @@ async function printSettingsActionTui(action: SettingsActionData): Promise<void>
 
 async function printPathsTui(paths: Record<string, unknown>, isBakinHome: unknown): Promise<void> {
   return renderInkReport(() => import('../src/core/cli/ui/readonly'), (m) => m.PathsReport, { paths, isBakinHome })
-}
-
-async function printAgentsListTui(agents: Array<{ id: string; name: string; status: string; model: string }>): Promise<void> {
-  return renderInkReport(() => import('../src/core/cli/ui/readonly'), (m) => m.AgentsListReport, { agents })
-}
-
-async function printAgentStatusTui(agentId: string, profile: Record<string, unknown>): Promise<void> {
-  return renderInkReport(() => import('../src/core/cli/ui/readonly'), (m) => m.AgentStatusReport, { agentId, profile })
-}
-
-async function printAgentTasksTui(agentId: string, tasks: Array<Record<string, unknown>>): Promise<void> {
-  return renderInkReport(() => import('../src/core/cli/ui/readonly'), (m) => m.AgentTasksReport, { agentId, tasks })
 }
 
 async function printPluginsListTui(plugins: Array<Record<string, unknown>>): Promise<void> {
@@ -106,28 +92,6 @@ async function printPluginRestoreResultTui(pluginId: string, result: PluginResto
   return renderInkReport(() => import('../src/core/cli/ui/readonly'), (m) => m.PluginRestoreResultReport, { pluginId, result })
 }
 
-async function printAgentPackagesListTui(agents: Array<Record<string, unknown>>): Promise<void> {
-  return renderInkReport(() => import('../src/core/cli/ui/readonly'), (m) => m.AgentPackagesListReport, { agents })
-}
-
-async function printAgentLessonsListTui(
-  agentId: string,
-  packageId: string,
-  lessons: Array<Record<string, unknown>>,
-): Promise<void> {
-  return renderInkReport(() => import('../src/core/cli/ui/readonly'), (m) => m.AgentLessonsListReport, { agentId, packageId, lessons })
-}
-
-async function printPackagesListTui(packages: Array<Record<string, unknown>>): Promise<void> {
-  return renderInkReport(() => import('../src/core/cli/ui/readonly'), (m) => m.PackagesListReport, { packages })
-}
-
-async function printPackageActionTui(actions: PackageActionData | PackageActionData[]): Promise<void> {
-  return renderInkReport(() => import('../src/core/cli/ui/readonly'), (m) => m.PackageActionReport, {
-    actions: Array.isArray(actions) ? actions : [actions],
-  })
-}
-
 async function printOnboardingCheckTui(result: CheckResult): Promise<void> {
   return renderInkReport(() => import('../src/core/cli/ui/onboarding'), (m) => m.OnboardingCheckReport, { result })
 }
@@ -143,56 +107,6 @@ async function printOnboardingInstallTui(result: InstallResult): Promise<void> {
 // ---------------------------------------------------------------------------
 // Commands
 // ---------------------------------------------------------------------------
-async function cmdAgentsList(opts: { json?: boolean } = {}): Promise<void> {
-  const result = await apiGet('/api/plugins/team/') as {
-    agents: Array<{ id: string; name: string; status: string; model: string }>
-  }
-  if (opts.json) {
-    print(result)
-    return
-  }
-  if (process.stdout.isTTY) {
-    await printAgentsListTui(result.agents)
-    return
-  }
-  for (const agent of result.agents) {
-    const statusIcon = agent.status === 'working' ? '●' : agent.status === 'online' ? '○' : '·'
-    console.log(`  ${statusIcon} ${agent.id}: ${agent.name} [${agent.status}] (${agent.model})`)
-  }
-}
-
-async function cmdAgentsSend(agentId: string, message: string): Promise<void> {
-  const result = await apiPost(`/api/agents/${agentId}/message`, { message })
-  if (process.stdout.isTTY) {
-    await printRuntimeActionTui({
-      action: 'message',
-      target: agentId,
-      detail: message,
-      result,
-    })
-    return
-  }
-  print(result)
-}
-
-async function cmdAgentsStatus(agentId: string, opts: { json?: boolean } = {}): Promise<void> {
-  const result = await apiGet(`/api/plugins/team/${agentId}`) as Record<string, unknown>
-  if (!opts.json && process.stdout.isTTY) {
-    await printAgentStatusTui(agentId, result)
-    return
-  }
-  print(result)
-}
-
-async function cmdAgentsTasks(agentId: string): Promise<void> {
-  const result = await apiGet(`/api/agents/${agentId}/tasks`) as { tasks: Array<Record<string, unknown>> }
-  if (process.stdout.isTTY) {
-    await printAgentTasksTui(agentId, result.tasks)
-    return
-  }
-  printTable(result.tasks, ['id', 'title', 'column'])
-}
-
 async function cmdSettingsGet(key?: string, opts: { json?: boolean } = {}): Promise<void> {
   const settings = await apiGet('/api/settings') as Record<string, unknown>
   if (key) {
@@ -780,335 +694,6 @@ async function cmdPluginsUnlink(pluginId: string, opts: { json?: boolean } = {})
     return
   }
   print(result)
-}
-
-// ---------------------------------------------------------------------------
-// Agent-package commands — `bakin agents ...`
-// ---------------------------------------------------------------------------
-
-interface AgentsCmdFlags {
-  adopt?: boolean
-  installAs?: string
-  replace?: boolean
-  keepBlocks?: boolean
-  deleteAgent?: boolean
-  check?: boolean
-  reclaim?: string[]
-  reclaimAll?: boolean
-  yes?: boolean
-  force?: boolean
-  json?: boolean
-}
-
-async function cmdAgentPackagesInstall(source: string, flags: AgentsCmdFlags): Promise<void> {
-  const body: Record<string, unknown> = { source }
-  if (flags.adopt) body.adopt = source // installer reads any truthy adopt as the agentId
-  if (flags.installAs) body.installAs = flags.installAs
-  if (flags.replace) body.replace = true
-  const result = await apiPost('/api/agent-packages/install', body)
-  if (!flags.json && process.stdout.isTTY) {
-    await printPackageActionTui({
-      action: 'installed',
-      scope: 'agent package',
-      target: flags.installAs ?? source,
-      result,
-    })
-    return
-  }
-  print(result)
-}
-
-async function cmdAgentPackagesList(flags: AgentsCmdFlags): Promise<void> {
-  const result = await apiGet('/api/agent-packages') as {
-    ok: boolean
-    agents: Array<{
-      agentId: string
-      state: string
-      version?: string
-      packageId?: string
-      entry?: { version?: string }
-    }>
-  }
-  if (flags.json) {
-    print(result)
-    return
-  }
-  if (process.stdout.isTTY) {
-    await printAgentPackagesListTui(result.agents)
-    return
-  }
-  console.log('Agents (package state):')
-  for (const a of result.agents) {
-    const version = a.version ?? a.entry?.version ?? '-'
-    const pkg = a.packageId ? `  [${a.packageId}]` : ''
-    console.log(`  ${a.agentId.padEnd(20)} ${a.state.padEnd(10)} ${version.padEnd(10)}${pkg}`)
-  }
-}
-
-async function cmdAgentPackagesRemove(agentId: string, flags: AgentsCmdFlags): Promise<void> {
-  const body: Record<string, unknown> = {}
-  if (flags.keepBlocks) body.keepBlocks = true
-  if (flags.deleteAgent) body.deleteAgent = true
-  if (flags.force) body.force = true
-  const result = await apiDelete(`/api/agent-packages/${encodeURIComponent(agentId)}`, body)
-  if (!flags.json && process.stdout.isTTY) {
-    await printPackageActionTui({
-      action: 'removed',
-      scope: 'agent package',
-      target: agentId,
-      result,
-    })
-    return
-  }
-  print(result)
-}
-
-async function cmdAgentPackagesOrphan(agentId: string, flags: AgentsCmdFlags): Promise<void> {
-  await cmdAgentPackagesRemove(agentId, { ...flags, deleteAgent: false })
-}
-
-async function cmdAgentPackagesDelete(agentId: string, flags: AgentsCmdFlags): Promise<void> {
-  await cmdAgentPackagesRemove(agentId, { ...flags, deleteAgent: true })
-}
-
-async function cmdAgentPackagesSync(agentId: string | undefined, flags: AgentsCmdFlags): Promise<void> {
-  const body: Record<string, unknown> = {}
-  if (flags.check) body.check = true
-  if (flags.reclaimAll) body.reclaim = 'all'
-  else if (flags.reclaim?.length) body.reclaim = flags.reclaim
-
-  const syncOne = async (id: string): Promise<unknown> => {
-    // apiPostJson (not apiPost) — the 409 migrationRequired response is a
-    // structured payload we branch on, not a transport error to throw.
-    const first = await apiPostJson(`/api/agent-packages/${encodeURIComponent(id)}/sync`, body)
-    let result = first.data as Record<string, unknown>
-    if (result.migrationRequired) {
-      const confirmed = flags.yes || await confirmPrompt(
-        `Package for "${id}" predates block-based projection. Run the one-time migration?\n` +
-        '  This FULLY OVERWRITES package-managed workspace files with freshly composed content\n' +
-        '  (a tarball backup is written to ~/.bakin/.backups/ first). Proceed?',
-      )
-      if (!confirmed) {
-        return { ok: false, error: 'Migration declined — agent left un-synced.' }
-      }
-      const migration = await apiPost('/api/agent-packages/migrate', {}) as Record<string, unknown>
-      if (!migration.ok) return migration
-      const second = await apiPostJson(`/api/agent-packages/${encodeURIComponent(id)}/sync`, body)
-      result = second.data as Record<string, unknown>
-    }
-    return result
-  }
-
-  const action = flags.check ? 'checked' : 'synced'
-
-  if (agentId) {
-    const result = await syncOne(agentId)
-    if (!flags.json && process.stdout.isTTY) {
-      await printPackageActionTui({ action, scope: 'agent', target: agentId, result })
-      return
-    }
-    print(result)
-    return
-  }
-
-  // No agentId — sync every runtime agent (managed agents get the full
-  // sequence; unmanaged agents get their context block maintained).
-  const list = await apiGet('/api/agent-packages') as {
-    agents: Array<{ agentId: string; state: string }>
-  }
-  const actions: PackageActionData[] = []
-  for (const a of list.agents) {
-    if (a.state === 'absent') continue
-    try {
-      const result = await syncOne(a.agentId)
-      if (!flags.json && process.stdout.isTTY) {
-        actions.push({ action, scope: 'agent', target: a.agentId, result })
-        continue
-      }
-      console.log(`${a.agentId}:`)
-      print(result)
-    } catch (err) {
-      if (!flags.json && process.stdout.isTTY) {
-        actions.push({
-          action,
-          scope: 'agent',
-          target: a.agentId,
-          result: { ok: false, error: err instanceof Error ? err.message : String(err) },
-        })
-        continue
-      }
-      console.error(`${a.agentId}: ${err instanceof Error ? err.message : String(err)}`)
-    }
-  }
-  if (!flags.json && process.stdout.isTTY) {
-    await printPackageActionTui(actions)
-  }
-}
-
-async function cmdAgentPackagesLessonsList(agentId: string): Promise<void> {
-  const result = await apiGet(`/api/agent-packages/${encodeURIComponent(agentId)}/lessons`) as {
-    ok: boolean
-    packageId: string
-    lessons: Array<{ lessonId: string; title: string; tags: string[]; enabled: boolean }>
-  }
-  if (process.stdout.isTTY) {
-    await printAgentLessonsListTui(agentId, result.packageId, result.lessons)
-    return
-  }
-  console.log(`Lessons for ${agentId} (package: ${result.packageId})`)
-  for (const l of result.lessons) {
-    const mark = l.enabled ? '[x]' : '[ ]'
-    const tags = l.tags.length > 0 ? ` (${l.tags.join(', ')})` : ''
-    console.log(`  ${mark} ${l.lessonId.padEnd(30)} ${l.title}${tags}`)
-  }
-}
-
-async function cmdAgentPackagesLessonsToggle(
-  agentId: string,
-  lessonId: string,
-  enabled: boolean,
-  opts: { json?: boolean } = {},
-): Promise<void> {
-  const result = await apiPost(
-    `/api/agent-packages/${encodeURIComponent(agentId)}/lessons/${encodeURIComponent(lessonId)}`,
-    { enabled },
-  )
-  if (!opts.json && process.stdout.isTTY) {
-    await printPackageActionTui({
-      action: enabled ? 'enabled' : 'disabled',
-      scope: 'lesson',
-      target: lessonId,
-      context: agentId,
-      result,
-    })
-    return
-  }
-  print(result)
-}
-
-// ---------------------------------------------------------------------------
-// Standalone-pack commands — `bakin packages ...`
-// ---------------------------------------------------------------------------
-
-async function cmdPackagesList(flags: AgentsCmdFlags): Promise<void> {
-  const result = await apiGet('/api/packages') as {
-    ok: boolean
-    packages: Array<{
-      id: string; kind: string; version: string; refCount: number; dependents: string[]
-    }>
-  }
-  const packages = result.packages.filter(p => p.kind !== 'agent')
-  if (flags.json) {
-    print({ ...result, packages })
-    return
-  }
-  if (process.stdout.isTTY) {
-    await printPackagesListTui(packages)
-    return
-  }
-  console.log('Installed packages:')
-  for (const p of packages) {
-    const refs = p.refCount > 0 ? `  (refCount=${p.refCount}: ${p.dependents.join(', ')})` : ''
-    console.log(`  ${p.id.padEnd(40)} ${p.kind.padEnd(15)} ${p.version}${refs}`)
-  }
-}
-
-async function cmdPackagesInstall(source: string, flags: AgentsCmdFlags): Promise<void> {
-  const body: Record<string, unknown> = { source }
-  if (flags.installAs) body.installAs = flags.installAs
-  if (flags.replace) body.replace = true
-  const result = await apiPost('/api/packages/install', body)
-  if (!flags.json && process.stdout.isTTY) {
-    await printPackageActionTui({
-      action: 'installed',
-      scope: 'package',
-      target: flags.installAs ?? source,
-      result,
-    })
-    return
-  }
-  print(result)
-}
-
-async function cmdPackagesRemove(packageId: string, flags: AgentsCmdFlags): Promise<void> {
-  const body: Record<string, unknown> = {}
-  if (flags.keepBlocks) body.keepBlocks = true
-  if (flags.force) body.force = true
-  const result = await apiDelete(`/api/packages/${encodeURIComponent(packageId)}`, body)
-  if (!flags.json && process.stdout.isTTY) {
-    await printPackageActionTui({
-      action: 'removed',
-      scope: 'package',
-      target: packageId,
-      result,
-    })
-    return
-  }
-  print(result)
-}
-
-async function cmdPackagesSync(packageId: string, flags: AgentsCmdFlags): Promise<void> {
-  const body: Record<string, unknown> = {}
-  if (flags.check) body.check = true
-  const result = await apiPost(`/api/packages/${encodeURIComponent(packageId)}/sync`, body)
-  if (!flags.json && process.stdout.isTTY) {
-    await printPackageActionTui({
-      action: flags.check ? 'checked' : 'synced',
-      scope: 'package',
-      target: packageId,
-      result,
-    })
-    return
-  }
-  print(result)
-}
-
-/** Parse a flag-style argv tail (everything after the positional args). */
-function parseAgentsFlags(args: string[]): AgentsCmdFlags {
-  const flags: AgentsCmdFlags = {}
-  for (let i = 0; i < args.length; i++) {
-    const a = args[i]
-    switch (a) {
-      case '--adopt':
-        flags.adopt = true
-        break
-      case '--install-as':
-        flags.installAs = args[++i]
-        break
-      case '--replace':
-        flags.replace = true
-        break
-      case '--keep-blocks':
-        flags.keepBlocks = true
-        break
-      case '--delete-agent':
-      case '--delete':
-        flags.deleteAgent = true
-        break
-      case '--orphan':
-        flags.deleteAgent = false
-        break
-      case '--check':
-        flags.check = true
-        break
-      case '--reclaim':
-        flags.reclaim = [...(flags.reclaim ?? []), args[++i]]
-        break
-      case '--reclaim-all':
-        flags.reclaimAll = true
-        break
-      case '--yes':
-        flags.yes = true
-        break
-      case '--force':
-        flags.force = true
-        break
-      case '--json':
-        flags.json = true
-        break
-    }
-  }
-  return flags
 }
 
 type CliDoctorResult = {
@@ -2999,54 +2584,7 @@ export async function main(): Promise<void> {
         break
 
       case 'agents':
-        if (sub === 'list') {
-          // `bakin agents list --packages` → package state view
-          // `bakin agents list`            → existing runtime view
-          if (args.includes('--packages')) {
-            await cmdAgentPackagesList(parseAgentsFlags(args.slice(2)))
-          } else {
-            await cmdAgentsList({ json: args.includes('--json') })
-          }
-        } else if (sub === 'status') {
-          if (!args[2]) await exitUsage('bakin agents status <id>')
-          await cmdAgentsStatus(args[2], { json: args.includes('--json') })
-        } else if (sub === 'tasks') {
-          if (!args[2]) await exitUsage('bakin agents tasks <id>')
-          await cmdAgentsTasks(args[2])
-        } else if (sub === 'send') {
-          if (!args[2] || !args[3]) await exitUsage('bakin agents send <id> <message>')
-          await cmdAgentsSend(args[2], args.slice(3).join(' '))
-        } else if (sub === 'install') {
-          if (!args[2]) await exitUsage('bakin agents install <path|github:user/repo[@ref][#subpath]> [--adopt] [--install-as <id>] [--replace]')
-          await cmdAgentPackagesInstall(args[2], parseAgentsFlags(args.slice(3)))
-        } else if (sub === 'orphan') {
-          if (!args[2]) await exitUsage('bakin agents orphan <agent-id> [--keep-blocks] [--force]')
-          await cmdAgentPackagesOrphan(args[2], parseAgentsFlags(args.slice(3)))
-        } else if (sub === 'delete') {
-          if (!args[2]) await exitUsage('bakin agents delete <agent-id> [--keep-blocks] [--force]')
-          await cmdAgentPackagesDelete(args[2], parseAgentsFlags(args.slice(3)))
-        } else if (sub === 'remove') {
-          if (!args[2]) await exitUsage('bakin agents remove <agent-id> [--keep-blocks] [--delete-agent] [--delete] [--force]')
-          await cmdAgentPackagesRemove(args[2], parseAgentsFlags(args.slice(3)))
-        } else if (sub === 'sync') {
-          // `bakin agents sync` (no id) syncs every agent; `bakin agents sync <id>` is targeted
-          const id = args[2] && !args[2].startsWith('--') ? args[2] : undefined
-          const flagsStart = id ? 3 : 2
-          await cmdAgentPackagesSync(id, parseAgentsFlags(args.slice(flagsStart)))
-        } else if (sub === 'lessons') {
-          const lessonSub = args[2]
-          if (lessonSub === 'list') {
-            if (!args[3]) await exitUsage('bakin agents lessons list <agent-id>')
-            await cmdAgentPackagesLessonsList(args[3])
-          } else if (lessonSub === 'enable' || lessonSub === 'disable') {
-            if (!args[3] || !args[4]) await exitUsage(`bakin agents lessons ${lessonSub} <agent-id> <lesson-id>`)
-            await cmdAgentPackagesLessonsToggle(args[3], args[4], lessonSub === 'enable', { json: args.includes('--json') })
-          } else {
-            await exitUnknownSubcommand('agents lessons', lessonSub, ['list', 'enable', 'disable'])
-          }
-        } else {
-          await exitUnknownSubcommand('agents', sub, ['list', 'status', 'tasks', 'send', 'install', 'orphan', 'delete', 'remove', 'update', 'lessons'])
-        }
+        await (await import('../src/cli/commands/agents')).run(args)
         break
 
       case 'settings':
@@ -3185,20 +2723,7 @@ export async function main(): Promise<void> {
         break
 
       case 'packages':
-        if (sub === 'install') {
-          if (!args[2]) await exitUsage('bakin packages install <path|github:user/repo[@ref][#subpath]> [--install-as <id>] [--replace]')
-          await cmdPackagesInstall(args[2], parseAgentsFlags(args.slice(3)))
-        } else if (sub === 'list') {
-          await cmdPackagesList(parseAgentsFlags(args.slice(2)))
-        } else if (sub === 'remove') {
-          if (!args[2]) await exitUsage('bakin packages remove <package-id> [--force] [--keep-blocks]')
-          await cmdPackagesRemove(args[2], parseAgentsFlags(args.slice(3)))
-        } else if (sub === 'sync') {
-          if (!args[2]) await exitUsage('bakin packages sync <package-id> [--check]')
-          await cmdPackagesSync(args[2], parseAgentsFlags(args.slice(3)))
-        } else {
-          await exitUnknownSubcommand('packages', sub, ['install', 'list', 'remove', 'sync'])
-        }
+        await (await import('../src/cli/commands/packages')).run(args)
         break
 
       case 'stop':
